+++ src/iteration_utilities/_utils.py
@@ -4,17 +4,16 @@
 import platform
 import sys
 
-__all__ = ['GE_PY36', 'GE_PY38', 'IS_CPYTHON', 'IS_PYPY', 'USES_VECTORCALL', '_default']
+__all__ = ['IS_CPYTHON', 'IS_CPYTHON_PY_3_12', 'IS_PYPY', 'USES_VECTORCALL', '_default']
 
 
-GE_PY36 = sys.version_info.major > 3 or (sys.version_info.major == 3 and
-                                         sys.version_info.minor >= 6)
-GE_PY38 = sys.version_info.major > 3 or (sys.version_info.major == 3 and
+_GE_PY38 = sys.version_info.major > 3 or (sys.version_info.major == 3 and
                                          sys.version_info.minor >= 8)
+IS_CPYTHON_PY_3_12 = sys.version_info.major == 3 and sys.version_info.minor == 12
 IS_PYPY = platform.python_implementation() == 'PyPy'
 IS_CPYTHON = platform.python_implementation() == 'CPython'
 
-USES_VECTORCALL = IS_CPYTHON and GE_PY38
+USES_VECTORCALL = IS_CPYTHON and _GE_PY38
 
 
 class _SentinelFactory: